+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+2.4.1b
+- Midnight: Fix stuff for new Midnight changes.
+- Add new color settings for Color Trinket instead of just green/red.
+- Fix Color Trinket on TBC/MoP not going green again when trinket is up.
+
 2.4.1
 - Add Jazggz profile (www.twitch.tv/jazggz). Thank you for sharing <3
 - The "Party Target Indicators" Widget now also has options to show who Enemy Arena Units are targeting on your PartyFrame. Also fixed for 5v5.